Validação dos dados do perfil de Usuario

+++ app/Models/User.php
@@ -59,17 +59,28 @@ class User extends Authenticatable implements JWTSubject
         return $user;
     }
 
+    /**
+     * @param array $data
+     * @return User
+     * @throws \Exception
+     */
     public function updateWithProfile(array $data) : User
     {
         try{
-            UserProfile::uploadPhoto($data['photo']);
+            if (isset($data['photo'])) {
+                UserProfile::uploadPhoto($data['photo']);
+            }
+
             \DB::beginTransaction();
             $this->fill($data);
             $this->save();
             UserProfile::saveProfile($this, $data);
             \DB::commit();
         } catch (\Exception $e) {
-            UserProfile::deleteFile($data['photo']);
+            if (isset($data['photo'])) {
+                UserProfile::deleteFile($data['photo']);
+            }
+
             \DB::rollBack();
             throw $e;
         }
